Re-added Package Control settings to repo.
I removed it before because it has a list of installed packages and
I saw that as being environment specific. But, the reality is that
whenever I move onto a different computer I want that list so I know
what I need to install.

Of course, maybe I should be putting my whole Packages directory in
version control…